refactor(canvas)
