+++ source/js/canvas/LoseScene.js
@@ -1,6 +1,6 @@
-import Animation from '../utils/animation';
-import Scene2D from '../utils/scene-2d.js';
-import easings from '../utils/easings.js';
+import Animation from './utils/Animation';
+import Scene2D from './utils/Scene2D';
+import easings from './utils/easings';
 
 const IMAGES_URLS = Object.freeze({
   key: `./img/module-4/lose-images/key.png`,
@@ -102,7 +102,7 @@ const LOCALS = Object.freeze({
   }
 });
 
-export default class CrocodileScene extends Scene2D {
+export default class LoseScene extends Scene2D {
   constructor() {
     const canvas = document.getElementById(`lose-scene-canvas`);
