+++ source/js/canvas/utils/Animation.js
