+++ source/js/canvas/utils/easings.js
